fix : Change integer to Double 양수를 생각했을떄 처음에 단순히 양의 부호인 수를 생각했습니다. 하지만 소수도 양수에 포함 되기 떄문에 기존에 integer를 실수인 Double로 변경

+++ src/main/java/calculator/Numbers.java
@@ -4,23 +4,23 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Numbers {
-    private List<Integer> numbers;
-    public Numbers(List<Integer> numbers) {
+    private List<Double> numbers;
+    public Numbers(List<Double> numbers) {
         validateNegative(numbers);
         this.numbers = numbers;
     }
 
-    private void validateNegative(List<Integer> numbers) {
-        for (Integer number : numbers) {
+    private void validateNegative(List<Double> numbers) {
+        for (Double number : numbers) {
             if (number < 0) {
                 throw new IllegalArgumentException("[ERROR] 양의 정수를 입력해주세요");
             }
         }
     }
 
-    public int getSum() {
-        int sum = 0;
-        for (Integer number : numbers) {
+    public Double getSum() {
+        Double sum = 0.0;
+        for (Double number : numbers) {
             sum += number;
         }
         return sum;
